==> notebooks/01_download_data.ipynb <==
## Tarbimise ja tootmise andmete allalaadimine

- loeme CSV failist tarbimise ja tootmise andmed.
- salvestame andmed faili `data/raw_consumption.json` ja `data/raw_production.json`

import pandas as pd
from pathlib import Path

project_root = Path("..").resolve()
data_dir = project_root / "data"
data_dir.mkdir(exist_ok=True)

Laeme andmes .csv failist.

csv_path = Path("..") /"data"/"electricity_production_and_consumption_10.12-11.12.csv"

df = pd.read_csv(csv_path, sep=";", encoding="latin1")
df.head()

Puhastame veerud.

numeric_cols = ["Tarbimine", "Tootmine", "Planeeritud tarbimine", "Planeeritud tootmine"]

for col in numeric_cols:
    df[col] = df[col].astype(str).str.replace(",", ".").astype(float)

df.head()

Lisame Timestampi

df["timestamp"] = pd.to_datetime(df["Ajatempel (UTC)"], unit="s")
df = df.sort_values("timestamp").reset_index(drop=True)

df[["Ajatempel (UTC)", "timestamp"]].head()

Koostame tarbimise JSON faili.

df_consumption = df[["timestamp", "Tarbimine"]].copy()
df_consumption.rename(columns={"Tarbimine": "consumption_mwh"}, inplace=True)

df_consumption.to_json(data_dir / "raw_consumption.json", orient="records", date_format="iso")

df_consumption.head()


Koostame tootmise JSON faili.

df_production = df[["timestamp", "Tootmine"]].copy()
df_production.rename(columns={"Tootmine": "production_mwh"}, inplace=True)

df_production.to_json(data_dir / "raw_production.json", orient="records", date_format="iso")

df_production.head()

Koostame planeeritud tarbimise JSON faili.

df_forecast = df[["timestamp", "Planeeritud tarbimine"]].copy()
df_forecast.rename(columns={"Planeeritud tarbimine": "forecast_consumption_mwh"}, inplace=True)

forecast_path = data_dir / "raw_forecast_consumption.json"
df_forecast.to_json(forecast_path, orient="records", date_format="iso")

forecast_path, df_forecast.head()

Koostame planeeritud tootmise JSON faili.

df_forecast = df[["timestamp", "Planeeritud tootmine"]].copy()
df_forecast.rename(columns={"Planeeritud tootmine": "forecast_production_mwh"}, inplace=True)

forecast_path = data_dir / "raw_forecast_production.json"
df_forecast.to_json(forecast_path, orient="records", date_format="iso")

forecast_path, df_forecast.head()

- Importisin tarbimise ja tootmise andmed CSV failist.
- Normaliseerisin numbriveerud.
- Teisendasin timestampi UNIX sekunditest kuupäevaks.
- Lõin JSON failid:
    - `raw_consumption.json`
    - `raw_production.json`
    - `raw_forecast_consumption.json`
    - `raw_forecast_production.json`
